requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
motor>=3.4.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
cryptography>=42.0.5
httpx>=0.27.0
pydantic-settings>=2.2.1
google-genai>=0.2.0
python-dotenv>=1.0.1
pytz>=2024.1
pytest>=8.1.1
pytest-asyncio>=0.23.6
python-jose[cryptography]>=3.3.0
email-validator>=2.1.1
